refactor: merge functions to `combine`
