--- src/context/index.ts
@@ -1,2 +1,0 @@
-export * from './createContext';
-export * from './mergeContext';
