+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from './helpers';
 export { FatcherError, isFatcherError } from './errors';
-export { mergeOptions } from './mergeOptions';
+import { combine } from './utils';
+export { combine };
 export * from './interfaces';
 export * from './fatcher';
 export * from './createScopedRequest';
