+++ src/utils.ts
@@ -1,3 +1,6 @@
+import { merge } from '@fatcherjs/utils-shared';
+import { Context, RequestOptions } from './interfaces';
+
 /**
  * Parse a url with `base` and `url`
  *
@@ -39,4 +42,33 @@ export function parseURL(base: string, url: string) {
     }
 
     return _url;
+}
+
+/**
+ * Merge headers into a headers
+ * @param headers
+ * @param patchHeaders
+ * @returns
+ */
+export function mergeHeaders(
+    headers: Headers | Record<string, string>,
+    patchHeaders: Headers | Record<string, string>
+) {
+    const h = headers instanceof Headers ? headers : new Headers();
+
+    for (const [key, value] of Object.entries(patchHeaders)) {
+        h.set(key, value);
+    }
+
+    return h;
+}
+
+export function combine<T extends RequestOptions | Context>(record: T, ...patches: Partial<T>[]): T {
+    return merge(record, patches, (merged, patch) => {
+        if (patch.headers) {
+            patch.headers = mergeHeaders(merged.headers || new Headers(), patch.headers);
+        }
+
+        return patch;
+    });
 }
